Unit test current_objects ModelManager.
