+++ common/tests/test_models_utils.py
@@ -7,8 +7,20 @@ from common.models.utils import get_current_transaction
 from common.models.utils import override_current_transaction
 from common.models.utils import set_current_transaction
 from common.tests import factories
+from common.tests.models import TestModel1
+from common.tests.models import model_with_history
 
 pytestmark = pytest.mark.django_db
+
+
+@pytest.fixture
+def test_model1_with_history(date_ranges):
+    return model_with_history(
+        factories.TestModel1Factory,
+        date_ranges,
+        version_group=factories.VersionGroupFactory.create(),
+        sid=1,
+    )
 
 
 def test_lazy_value():
@@ -44,3 +56,9 @@ def test_override_current_transaction():
         assert get_current_transaction() is tx2
 
     assert get_current_transaction() is tx
+
+
+def test_current_objects_model_manager(test_model1_with_history):
+    for model_version in test_model1_with_history.all_models:
+        set_current_transaction(model_version.transaction)
+        assert TestModel1.current_objects.count() == 1
